Minor updates to projects.c
Updated some of the functions in projects.c, changed a malloc() call to calloc(). 

Ensured that in task_add_interactive() that user-provided variables are zero'd 
before the next loop iteration.

Added initial code for some new functions as well as comments for when I have time 
to come back to them and flesh them out further.