Minor Updates to projects.c
Made some minor changes to projects.c, exploring the idea of treating each 
query as a struct instead of a few random vars. This may end up as a
double linked list stored in memory when read from the database. 
I'm unsure if this is in any way superior, but it should at 
least be clearer what's being worked on, and the whole query 
can be passed to a function at once.